+++ doc/05_acting/06_main_frame_publisher.md
@@ -12,15 +12,13 @@ Julian Graf
 
 29.03.2023
 
----
-
 <!-- TOC -->
-* [Main frame publisher](#main-frame-publisher)
-  * [Author](#author)
-  * [Date](#date)
-  * [Main frame publisher](#main-frame-publisher-1)
-  * [How to use](#how-to-use)
-  * [Known issues](#known-issues)
+- [Main frame publisher](#main-frame-publisher)
+  - [Author](#author)
+  - [Date](#date)
+  - [Main frame publisher](#main-frame-publisher-1)
+  - [How to use](#how-to-use)
+  - [Known issues](#known-issues)
 <!-- TOC -->
 
 ## Main frame publisher
